requirements.txt
stomp.py
mysql-connector-python
python-etcd
flask
